Adaptation to Vaadin 23 done (#17)
