+++ src/main/java/io/kyberorg/example/ExampleApplication.java
@@ -1,7 +1,16 @@
 package io.kyberorg.example;
 
+import com.vaadin.flow.component.page.AppShellConfigurator;
+import com.vaadin.flow.component.page.Inline;
+import com.vaadin.flow.component.page.Push;
+import com.vaadin.flow.component.page.Viewport;
+import com.vaadin.flow.server.AppShellSettings;
+import com.vaadin.flow.server.PWA;
+import com.vaadin.flow.theme.Theme;
+import com.vaadin.flow.theme.lumo.Lumo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
 /**
  * Example Application.
@@ -9,7 +18,16 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * @since 1.0
  */
 @SpringBootApplication
-public class ExampleApplication {
+@Push
+@Viewport("width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes")
+@PWA(
+        name = "Example App",
+        shortName = "example",
+        offlinePath = "offline-page.html",
+        offlineResources = {"images/logo.png"},
+        description = "Simple Spring Boot Application with REST, Vaadin and Database ")
+@Theme(themeClass = Lumo.class, variant = Lumo.DARK)
+public class ExampleApplication extends SpringBootServletInitializer implements AppShellConfigurator {
 
     /**
      * Start Point.
@@ -20,4 +38,20 @@ public class ExampleApplication {
         SpringApplication.run(ExampleApplication.class, args);
     }
 
+    @Override
+    public void configurePage(final AppShellSettings settings) {
+        settings.addFavIcon("icon", "/icons/favicon-32x32.png", "32x32");
+        settings.addLink("shortcut icon", "/icons/favicon-16x16.png");
+        settings.addLink("apple-touch-icon", "/icons/apple-touch-icon.png");
+        settings.addLink("manifest", "/site.webmanifest");
+        settings.addLink("mask-icon", "/icons/safari-pinned-tab.svg");
+
+        settings.addMetaTag("apple-mobile-web-app-title", "example-app");
+        settings.addMetaTag("application-name", "example-app");
+        settings.addMetaTag("msapplication-TileColor", "#ffc40d");
+        settings.addMetaTag("theme-color", "#000000");
+
+        //Splash (Loading) Screen
+        settings.addInlineFromFile("splash-screen.html", Inline.Wrapping.NONE);
+    }
 }
