+++ src/main/java/io/kyberorg/example/ui/MainView.java
@@ -8,21 +8,12 @@ import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.html.Image;
 import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.page.AppShellConfigurator;
-import com.vaadin.flow.component.page.Inline;
-import com.vaadin.flow.component.page.Push;
-import com.vaadin.flow.component.page.Viewport;
 import com.vaadin.flow.component.tabs.Tab;
 import com.vaadin.flow.component.tabs.Tabs;
 import com.vaadin.flow.router.BeforeEnterEvent;
 import com.vaadin.flow.router.BeforeEnterObserver;
 import com.vaadin.flow.router.HighlightConditions;
 import com.vaadin.flow.router.RouterLink;
-import com.vaadin.flow.server.AppShellSettings;
-import com.vaadin.flow.server.PWA;
-import com.vaadin.flow.spring.annotation.UIScope;
-import com.vaadin.flow.theme.Theme;
-import com.vaadin.flow.theme.lumo.Lumo;
 import io.kyberorg.example.App;
 import io.kyberorg.example.util.AppUtils;
 
@@ -35,18 +26,9 @@ import java.util.Map;
  * @since 1.0
  */
 //@SpringComponent
-@UIScope
-@Push
-@Viewport("width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes")
-@PWA(
-        name = "Example App",
-        shortName = "example",
-        offlinePath = "offline-page.html",
-        offlineResources = {"images/logo.png"},
-        description = "Simple Spring Boot Application with REST, Vaadin and Database ")
-@Theme(themeClass = Lumo.class, variant = Lumo.DARK)
+//@UIScope
 @CssImport("./css/main_view.css")
-public class MainView extends AppLayout implements BeforeEnterObserver, AppShellConfigurator {
+public class MainView extends AppLayout implements BeforeEnterObserver {
 
     private final Tabs tabs = new Tabs();
     private final Map<Class<? extends Component>, Tab> tabToTarget = new HashMap<>();
@@ -115,23 +97,6 @@ public class MainView extends AppLayout implements BeforeEnterObserver, AppShell
     @Override
     public void beforeEnter(final BeforeEnterEvent beforeEnterEvent) {
         tabs.setSelectedTab(tabToTarget.get(beforeEnterEvent.getNavigationTarget()));
-    }
-
-    @Override
-    public void configurePage(final AppShellSettings settings) {
-        settings.addFavIcon("icon", "/icons/favicon-32x32.png", "32x32");
-        settings.addLink("shortcut icon", "/icons/favicon-16x16.png");
-        settings.addLink("apple-touch-icon", "/icons/apple-touch-icon.png");
-        settings.addLink("manifest", "/site.webmanifest");
-        settings.addLink("mask-icon", "/icons/safari-pinned-tab.svg");
-
-        settings.addMetaTag("apple-mobile-web-app-title", "example-app");
-        settings.addMetaTag("application-name", "example-app");
-        settings.addMetaTag("msapplication-TileColor", "#ffc40d");
-        settings.addMetaTag("theme-color", "#000000");
-
-        //Splash (Loading) Screen
-        settings.addInlineFromFile("splash-screen.html", Inline.Wrapping.NONE);
     }
 
     public static class IDs {
